+++ lib/src/controller.dart
@@ -8,5 +8,8 @@ export 'package:mvc_application/controller.dart' hide AppController, runApp;
 /// The Controller for the App itself
 export 'package:playhouse/src/app/controller/app_controller.dart';
 
+/// Controls the Authentication into the App
+export 'package:playhouse/src/app/controller/initAsync/playhouse_signin.dart';
+
 /// Authentication
 export 'package:auth/auth.dart';