+++ lib/src/home/view/tabview/modules/scrapbook_modules_design_screen.dart
@@ -1,0 +1,11 @@
+// Copyright 2021 Grey & Ivy Inc. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// The UI code
+import 'package:playhouse/src/view.dart';
+
+class ScrapbookDesignScreen extends ModuleTypeScreen {
+  const ScrapbookDesignScreen({Key key}):super(key: key);
+}
+