+++ lib/src/view.dart
@@ -4,15 +4,34 @@
 
 
 /// MVC framework
-export 'package:mvc_application/view.dart' hide Card, runApp;
+export 'package:mvc_application/view.dart' hide runApp;
 
 /// https://pub.dev/packages/cached_network_image/
 export 'package:cached_network_image/cached_network_image.dart';
 
-///  https://pub.dev/packages/scrollytell
-export 'package:scrollytell/scrollytell.dart';
+/// Scrapbook main screen
+export 'package:playhouse/src/home/view/scrapbook_main.dart'; //  hide ScrapBookApp
 
-export 'package:playhouse/src/home/view/home_page.dart';
+/// Scrapbook AppBar
+export 'package:playhouse/src/home/view/appbar/scrapbook_appbar.dart';
 
-export 'package:playhouse/src/home/view/scrapbook.dart';
+/// Scrapbook Drawer
+export 'package:playhouse/src/home/view/drawer/scrapbook_drawer.dart';
 
+/// Scrapbook Design screen
+export 'package:playhouse/src/home/view/tabview/modules/scrapbook_modules_design_screen.dart';
+
+/// Scrapbook Build screen
+export 'package:playhouse/src/home/view/tabview/modules/scrapbook_modules_build_screen.dart';
+
+/// Scrapbook Submodules screen
+export 'package:playhouse/src/home/view/tabview/submodule/scrapbook_submodules.dart';
+
+/// Modules screen (Parent of the two screens above: Design & Build
+export 'package:playhouse/src/home/view/tabview/modules/scrapbook_modules_type_screen.dart';
+
+/// Individual Module screen
+export 'package:playhouse/src/home/view/tabview/modules/scrapbook_module_screen.dart';
+
+/// Individual Task screen
+export 'package:playhouse/src/home/view/tabview/tasks/scrapbook_tasks.dart'; // hide Card;